Various changes I've been working on.
No new functionality. Renamed Task.cs to SCTask.cs, renamed Assembly.cs to ExecAssembly.cs, added Spawn.cs. Fixed references in SaltedCaramelTests.

+++ SaltedCaramel/Win32/kernel32.cs
@@ -57,8 +57,8 @@ namespace SaltedCaramel.Win32
 
         [DllImport("kernel32.dll", SetLastError = true, ExactSpelling = true)]
         internal static extern IntPtr VirtualAlloc(
-            uint lpAddress, 
-            uint dwSize, 
+            IntPtr lpAddress, 
+            IntPtr dwSize, 
             AllocationType flAllocationType, 
             MemoryProtection flProtect);
